refactor: split workspace settings into 3 tabs (General/Instruction/Members)
- General: workspace name/slug editing + danger zone (owner only)
- Instruction: global instruction editor with Cmd+S (all members)
- Members: invites + member list (unchanged)
